requirements.txt
autopep8==2.0.2
flake8==6.0.0
isort==5.12.0
Jinja2==3.1.2
pandas==2.0.1
psycopg2==2.9.6
python-dotenv==1.0.0
